# codice/Server/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "API of snm project",
    "description": "I dont know what else to write here",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/caricaGeneri": {
      "get": {
        "description": "restituisce i generi musicali",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/registrazione": {
      "post": {
        "description": "inserisce un nuovo utente nel database",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nomeUtente": {
                  "example": "Sir"
                },
                "password": {
                  "example": "password"
                },
                "email": {
                  "example": "luca.salvadori@example.it"
                },
                "nome": {
                  "example": "Luca"
                },
                "cognome": {
                  "example": "Salvadori"
                },
                "sesso": {
                  "example": "M"
                },
                "dataNascita": {
                  "example": "1998-01-01"
                },
                "generi": {
                  "example": ["Rock", "Pop"]
                },
                "artistaPreferito": {
                  "example": "Queen"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Documento creato"
          },
          "400": {
            "description": "Dati non validi (uno o più)"
          },
          "500": {
            "description": "Errore del database"
          }
        }
      }
    },
    "/login": {
      "post": {
        "description": "effettua il login",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "luca.salvadori@example.it"
                },
                "password": {
                  "example": "password"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login effettuato"
          },
          "400": {
            "description": "Dati non validi"
          },
          "401": {
            "description": "Credenziali non valide"
          },
          "500": {
            "description": "Errore del database"
          }
        }
      }
    },
    "/user/password": {
      "post": {
        "description": "modifica la password dell'utente",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "data",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "example": "df2s4s6rty889dsce53sde45"
                },
                "oldPassword": {
                  "example": "password"
                },
                "newPassword": {
                  "example": "newPassword"
                }
              }
            }
          }
        ],
        "responses": {
          "200":{ 
            "description": "Password modificata"
          },
          "400": {
            "description": "Dati non validi"
          },
          "401": {
            "description": "Non autorizzato"
          },
          "403": {
            "description": "Non autorizzato a modificare la password"
          },
          "404" : {
            "description": "Utente non trovato o password non corretta"
          },
          "500": {
            "description": "Errore del database"
          }
        }
      }
    },
    "/logout": {
      "post": {
        "description": "effettua il logout",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/user/{id}/playlist": {
      "get": {
        "description": "restituisce le playlist di un utente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK o vodi"
          },
          "400": {
            "description": "Id non valido"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Utente non trovato"
          },
          "500": {
            "description": "Errore del database"
          }
        }
      },
      "put": {
        "description": "aggiugne o rimuove una playlist dall'utente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "idplaylist",
            "in": "body",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Playlist aggiunta o rimossa"
          },
          "400": {
            "description": "Id utente o playlist non valido"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Non autorizzato a modificare la playlist"
          },
          "404": {
            "description": "Utente o playlist non trovati"
          },
          "500": {
            "description": "Errore del database"
          }
        }
      }
    },
    "/playlist": {
      "get": {
        "description": "restituisce un numero di  playlist pubbliche",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Errore del database"
          }
        }
      },
      "post": {
        "description": "crea una nuova playlist",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "data",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "creatore": {
                  "example": "df2s4s6rty889dsce53sde45"
                },
                "pubblica": {
                  "example": "true"
                },
                "titolo": {
                  "example": "Playlist di Luca"
                },
                "descrizione": {
                  "example": "Una playlist di Luca"
                },
                "tag": {
                  "example": ["Rock", "Pop", "Queen"]
                },
                "canzoni": {
                  "example": ["df2s4s6rty889dsce53sde45", "df2s4s6rty889dsce53sde46"]
                },
                "durata": {
                  "example": "148060"
                },
                "autore": {
                  "example": "df2s4s6rty889dsce53sde45"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Playlist creata"
          },
          "400": {
            "description": "Dati non validi"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Errore del database"
          }
        }
      }
    },
    "/playlist/{id}": {
      "get": {
        "description": "restiuisce la playlist con id specificato",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Id non valido"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Playlist non trovata"
          },
          "500": {
            "description": "Errore del database"
          }
        }
      },
      "delete": {
        "description": "Elimina la playlist con id specificato",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "idRichiedente": {
                  "example": "df2s4s6rty889dsce53sde45"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Playlist eliminata"
          },
          "400": {
            "description": "Id o id del richiedente non valido"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Non autorizzato a eliminare la playlist"
          },
          "404": {
            "description": "Playlist o utente non trovato"
          },
          "500": {
            "description": "Errore del database"
          }
        }
      }
    },
    "/user": {
      "get": {
        "description": "restituisce una lista di utenti",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Errore del database"
          }
        }
      }
    },
    "/user/{id}": {
      "get": {
        "description": "restituisce un utente specifico",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Id non valido"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Utente non trovato"
          },
          "500": {
            "description": "Errore del database"
          }
        }
      },
      "put": {
        "description": "modifica i dati personali di un utente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "data",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "example": "Luca"
                },
                "cognome": {
                  "example": "Salvadori"
                },
                "sesso": {
                  "example": "M"
                },
                "dataNascita": {
                  "example": "1998-01-01"
                },
                "generi": {
                  "example": ["Rock", "Pop"]
                },
                "artistaPreferito": {
                  "example": "Queen"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Utente modificato"
          },
          "400": {
            "description": "Dati o id utente non validi"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Non autorizzato a modificare l'utente"
          },
          "404": {
            "description": "Utente non trovato"
          },
          "500": {
            "description": "Errore del database"
          }
        }
      },
      "delete": {
        "description": "elimina un utente dal sistema",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "example": "hfdskkdsfhe8"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Utente eliminato"
          },
          "400": {
            "description": "Password o id utente non validi"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Non autorizzato a eliminare l'utente"
          },
          "404": {
            "description": "Utente non trovato o password non corretta"
          },
          "500": {
            "description": "Errore del database"
          }
        }
      }
    },
    "/user/{id}/users": {
      "get": {
        "description": "restituisce la lista di amici di un utente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK oppure void"
          },
          "400": {
            "description": "Id non valido"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Utente non trovato"
          },
          "500": {
            "description": "Errore del database"
          }
        }
      },
      "put": {
        "description": "aggiunge oppure toglie un follower all'utente",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "example": "df2s4s6rty889dsce53sde45"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Follower aggiunto o rimosso"
          },
          "400": {
            "description": "Id utente o follower non valido"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Non autorizzato a modificare i follower"
          },
          "404": {
            "description": "Utente o follower non trovati"
          },
          "500": {
            "description": "Errore del database"
          }
        }
      }
    },
    "/song/{id}": {
      "get": {
        "description": "restituisce una canzone con id specificato",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "type",
            "in": "query",
            "type": "string",
            "example": "spotifyId or idDb"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Id o tipo non valido"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500":{
            "description": "Errore del database"
          }
        }
      }
    },
    "/song": {
      "post": {
        "description": "restituisce una lista di canzoni",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "idCanzoni": {
                  "example": "[dzf2s4s6rty889dsce53sde45, dzf2s4s6rty889dsce53sde46]"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Id non valido"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Errore del database"
          }
        }
      }
    },
    "/playlist/{id}/modifica": {
      "put": {
        "description": "modifica i dati di una playlist",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "data",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "titolo": {
                  "example": "Playlist di Luca"
                },
                "descrizione": {
                  "example": "Una playlist di Luca"
                },
                "tag": {
                  "example": ["Rock", "Pop", "Queen"]
                },
                "canzoni": {
                  "example": ["df2s4s6rty889dsce53sde45", "df2s4s6rty889dsce53sde46"]
                },
                "durata": {
                  "example": "148060"
                },
                "autore": {
                  "example": "df2s4s6rty889dsce53sde45"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Playlist modificata"
          },
          "400": {
            "description": "Dati o id playlist non validi"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Playlist non trovata"
          },
          "500": {
            "description": "Errore del database"
          }
        }
      }
    },
    "/playlist/{id}/songs": {
      "put": {
        "description": "aggiunge oppure toglie una canzone dalla playlist",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "idCanzone": {
                  "example": "dzf2s4s6rty889dsce53sde45"
                },
                "durata": {
                  "example": "259630"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Canzone aggiunta o rimossa"
          },
          "400": {
            "description": "Durata. id canzone o id playlist non valida"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Canzone o playlist non trovata, oppure errore nel calcolo della durata"
          },
          "500": {
            "description": "Errore del database"
          }
        }
      }
    },
    "/search": {
      "get": {
        "description": "ricerca canzoni, playlist e/o utenti in base alla tipologia di ricerca richiesta",
        "parameters": [
          {
            "name": "input",
            "in": "query",
            "type": "string"
          },
          {
            "name": "type",
            "in": "query",
            "type": "string",
            "example": "song o all"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Tipo di ricerca o input non valido"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Nessun risultato trovato oppure errore nella ricerca della canzone da spotify"
          },
          "500": {
            "description": "Errore del database"
          }
        }
      }
    }
  }
}